requirements.txt
icrawler
six
pillow
requests
bs4
lxml
datasets
transformers[torch]
evaluate
torchvision
scikit-learn
tensorboard